Add publish.md with new release instructions
The new file publish.md has been added to provide detailed instructions for publishing a new release. This includes updating the version number in the `main.tf` file and creating a new tag with the version number. The addition of this file will make the release process more clear and straight forward.